small code adjustment
