+++ pkg/manager/member/pd_member_manager.go
@@ -85,10 +85,6 @@ func (pmm *pdMemberManager) syncPDServiceForTidbCluster(tc *v1alpha1.TidbCluster
 	newSvc := pmm.getNewPDServiceForTidbCluster(tc)
 	oldSvc, err := pmm.svcLister.Services(ns).Get(controller.PDMemberName(tcName))
 	if errors.IsNotFound(err) {
-		err = SetServiceLastAppliedConfigAnnotation(newSvc)
-		if err != nil {
-			return err
-		}
 		return pmm.svcControl.CreateService(tc, newSvc)
 	}
 	if err != nil {
@@ -104,10 +100,6 @@ func (pmm *pdMemberManager) syncPDServiceForTidbCluster(tc *v1alpha1.TidbCluster
 		svc.Spec = newSvc.Spec
 		// TODO add unit test
 		svc.Spec.ClusterIP = oldSvc.Spec.ClusterIP
-		err = SetServiceLastAppliedConfigAnnotation(&svc)
-		if err != nil {
-			return err
-		}
 		return pmm.svcControl.UpdateService(tc, &svc)
 	}
 
@@ -121,10 +113,6 @@ func (pmm *pdMemberManager) syncPDHeadlessServiceForTidbCluster(tc *v1alpha1.Tid
 	newSvc := pmm.getNewPDHeadlessServiceForTidbCluster(tc)
 	oldSvc, err := pmm.svcLister.Services(ns).Get(controller.PDPeerMemberName(tcName))
 	if errors.IsNotFound(err) {
-		err = SetServiceLastAppliedConfigAnnotation(newSvc)
-		if err != nil {
-			return err
-		}
 		return pmm.svcControl.CreateService(tc, newSvc)
 	}
 	if err != nil {
@@ -140,10 +128,6 @@ func (pmm *pdMemberManager) syncPDHeadlessServiceForTidbCluster(tc *v1alpha1.Tid
 		svc.Spec = newSvc.Spec
 		// TODO add unit test
 		svc.Spec.ClusterIP = oldSvc.Spec.ClusterIP
-		err = SetServiceLastAppliedConfigAnnotation(newSvc)
-		if err != nil {
-			return err
-		}
 		return pmm.svcControl.UpdateService(tc, &svc)
 	}
 
@@ -164,11 +148,7 @@ func (pmm *pdMemberManager) syncPDStatefulSetForTidbCluster(tc *v1alpha1.TidbClu
 		return err
 	}
 	if errors.IsNotFound(err) {
-		newPDSet.Spec.Replicas = func()*int32 { var i int32 = 1; return &i }();
-		err = SetLastAppliedConfigAnnotation(newPDSet)
-		if err != nil {
-			return err
-		}
+		newPDSet.Spec.Replicas = func() *int32 { var i int32 = 1; return &i }()
 		if err := pmm.setControl.CreateStatefulSet(tc, newPDSet); err != nil {
 			return err
 		}
@@ -204,10 +184,6 @@ func (pmm *pdMemberManager) syncPDStatefulSetForTidbCluster(tc *v1alpha1.TidbClu
 		set := *oldPDSet
 		set.Spec.Template = newPDSet.Spec.Template
 		*set.Spec.Replicas = *newPDSet.Spec.Replicas
-		err = SetLastAppliedConfigAnnotation(&set)
-		if err != nil {
-			return err
-		}
 		return pmm.setControl.UpdateStatefulSet(tc, &set)
 	}
