+++ pkg/manager/member/tikv_member_manager.go
@@ -105,10 +105,6 @@ func (tkmm *tikvMemberManager) syncServiceForTidbCluster(tc *v1alpha1.TidbCluste
 	newSvc := tkmm.getNewServiceForTidbCluster(tc, svcConfig)
 	oldSvc, err := tkmm.svcLister.Services(ns).Get(svcConfig.MemberName(tcName))
 	if errors.IsNotFound(err) {
-		err = SetServiceLastAppliedConfigAnnotation(newSvc)
-		if err != nil {
-			return err
-		}
 		return tkmm.svcControl.CreateService(tc, newSvc)
 	}
 	if err != nil {
@@ -124,10 +120,6 @@ func (tkmm *tikvMemberManager) syncServiceForTidbCluster(tc *v1alpha1.TidbCluste
 		svc.Spec = newSvc.Spec
 		// TODO add unit test
 		svc.Spec.ClusterIP = oldSvc.Spec.ClusterIP
-		err = SetServiceLastAppliedConfigAnnotation(newSvc)
-		if err != nil {
-			return err
-		}
 		return tkmm.svcControl.UpdateService(tc, &svc)
 	}
 
@@ -148,10 +140,6 @@ func (tkmm *tikvMemberManager) syncStatefulSetForTidbCluster(tc *v1alpha1.TidbCl
 		return err
 	}
 	if errors.IsNotFound(err) {
-		err = SetLastAppliedConfigAnnotation(newSet)
-		if err != nil {
-			return err
-		}
 		err = tkmm.setControl.CreateStatefulSet(tc, newSet)
 		if err != nil {
 			return err
@@ -188,10 +176,6 @@ func (tkmm *tikvMemberManager) syncStatefulSetForTidbCluster(tc *v1alpha1.TidbCl
 		set := *oldSet
 		set.Spec.Template = newSet.Spec.Template
 		*set.Spec.Replicas = *newSet.Spec.Replicas
-		err = SetLastAppliedConfigAnnotation(&set)
-		if err != nil {
-			return err
-		}
 		return tkmm.setControl.UpdateStatefulSet(tc, &set)
 	}
