+++ pkg/manager/member/utils.go
@@ -18,15 +18,11 @@ import (
 	"fmt"
 
 	"github.com/golang/glog"
+	"github.com/pingcap/tidb-operator/pkg/controller"
 	apps "k8s.io/api/apps/v1beta1"
 	corev1 "k8s.io/api/core/v1"
 
 	apiequality "k8s.io/apimachinery/pkg/api/equality"
-)
-
-const (
-	// LastAppliedConfigAnnotation is annotation key of last applied configuration
-	LastAppliedConfigAnnotation = "pingcap.com/last-applied-configuration"
 )
 
 func timezoneMountVolume() (corev1.VolumeMount, corev1.Volume) {
@@ -60,31 +56,9 @@ func statefulSetInNormal(set *apps.StatefulSet) bool {
 	return set.Status.CurrentRevision == set.Status.UpdateRevision && set.Generation <= *set.Status.ObservedGeneration
 }
 
-// SetLastAppliedConfigAnnotation set last applied config info to Statefulset's annotation
-func SetLastAppliedConfigAnnotation(set *apps.StatefulSet) error {
-	setApply, err := encode(set.Spec)
-	if err != nil {
-		return err
-	}
-	if set.Annotations == nil {
-		set.Annotations = map[string]string{}
-	}
-	set.Annotations[LastAppliedConfigAnnotation] = setApply
-
-	templateApply, err := encode(set.Spec.Template.Spec)
-	if err != nil {
-		return err
-	}
-	if set.Spec.Template.Annotations == nil {
-		set.Spec.Template.Annotations = map[string]string{}
-	}
-	set.Spec.Template.Annotations[LastAppliedConfigAnnotation] = templateApply
-	return nil
-}
-
 // GetLastAppliedConfig get last applied config info from Statefulset's annotation
 func GetLastAppliedConfig(set *apps.StatefulSet) (*apps.StatefulSetSpec, *corev1.PodSpec, error) {
-	specAppliedConfig, ok := set.Annotations[LastAppliedConfigAnnotation]
+	specAppliedConfig, ok := set.Annotations[controller.LastAppliedConfigAnnotation]
 	if !ok {
 		return nil, nil, fmt.Errorf("statefulset:[%s/%s] not found spec's apply config", set.GetNamespace(), set.GetName())
 	}
@@ -94,7 +68,7 @@ func GetLastAppliedConfig(set *apps.StatefulSet) (*apps.StatefulSetSpec, *corev1
 		return nil, nil, err
 	}
 
-	podSpecAppliedConfig, ok := set.Spec.Template.Annotations[LastAppliedConfigAnnotation]
+	podSpecAppliedConfig, ok := set.Spec.Template.Annotations[controller.LastAppliedConfigAnnotation]
 	if !ok {
 		return nil, nil, fmt.Errorf("statefulset:[%s/%s] not found template spec's apply config", set.GetNamespace(), set.GetName())
 	}
@@ -107,18 +81,10 @@ func GetLastAppliedConfig(set *apps.StatefulSet) (*apps.StatefulSetSpec, *corev1
 	return spec, podSpec, nil
 }
 
-func encode(obj interface{}) (string, error) {
-	b, err := json.Marshal(obj)
-	if err != nil {
-		return "", err
-	}
-	return string(b), nil
-}
-
 // EqualStatefulSet compare the new Statefulset's spec with old Statefulset's last applied config
 func EqualStatefulSet(new apps.StatefulSet, old apps.StatefulSet) (bool, error) {
 	oldConfig := apps.StatefulSetSpec{}
-	if lastAppliedConfig, ok := old.Annotations[LastAppliedConfigAnnotation]; ok {
+	if lastAppliedConfig, ok := old.Annotations[controller.LastAppliedConfigAnnotation]; ok {
 		err := json.Unmarshal([]byte(lastAppliedConfig), &oldConfig)
 		if err != nil {
 			glog.Errorf("unmarshal Statefulset: [%s/%s]'s applied config failed,error: %v", old.GetNamespace(), old.GetName(), err)
@@ -132,7 +98,7 @@ func EqualStatefulSet(new apps.StatefulSet, old apps.StatefulSet) (bool, error) 
 // EqualTemplate compare the new podTemplateSpec's spec with old podTemplateSpec's last applied config
 func EqualTemplate(new corev1.PodTemplateSpec, old corev1.PodTemplateSpec) (bool, error) {
 	oldConfig := corev1.PodSpec{}
-	if lastAppliedConfig, ok := old.Annotations[LastAppliedConfigAnnotation]; ok {
+	if lastAppliedConfig, ok := old.Annotations[controller.LastAppliedConfigAnnotation]; ok {
 		err := json.Unmarshal([]byte(lastAppliedConfig), &oldConfig)
 		if err != nil {
 			glog.Errorf("unmarshal PodTemplate: [%s/%s]'s applied config failed,error: %v", old.GetNamespace(), old.GetName(), err)
@@ -143,23 +109,10 @@ func EqualTemplate(new corev1.PodTemplateSpec, old corev1.PodTemplateSpec) (bool
 	return false, nil
 }
 
-// SetServiceLastAppliedConfigAnnotation set last applied config info to Service's annotation
-func SetServiceLastAppliedConfigAnnotation(svc *corev1.Service) error {
-	svcApply, err := encode(svc.Spec)
-	if err != nil {
-		return err
-	}
-	if svc.Annotations == nil {
-		svc.Annotations = map[string]string{}
-	}
-	svc.Annotations[LastAppliedConfigAnnotation] = svcApply
-	return nil
-}
-
 // EqualService compare the new Service's spec with old Service's last applied config
 func EqualService(new, old *corev1.Service) (bool, error) {
 	oldSpec := corev1.ServiceSpec{}
-	if lastAppliedConfig, ok := old.Annotations[LastAppliedConfigAnnotation]; ok {
+	if lastAppliedConfig, ok := old.Annotations[controller.LastAppliedConfigAnnotation]; ok {
 		err := json.Unmarshal([]byte(lastAppliedConfig), &oldSpec)
 		if err != nil {
 			glog.Errorf("unmarshal ServiceSpec: [%s/%s]'s applied config failed,error: %v", old.GetNamespace(), old.GetName(), err)
